% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{acquireTSS}
\alias{acquireTSS}
\title{acquireTSS
Retrieves all tss data from a given TSS experiment.}
\usage{
acquireTSS(expName, tssNum)
}
\arguments{
\item{expName}{an object of class tssExp containing the tss data}

\item{tssNum}{the slot number of the tss data to be clustered}
}
\value{
a list object containing TSS data from the entire dataset
}
\description{
acquireTSS
Retrieves all tss data from a given TSS experiment.
}

