% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAnnotationToTSR.R
\name{addAnnotationToTSR}
\alias{addAnnotationToTSR}
\title{addAnnotationToTSR}
\usage{
addAnnotationToTSR(experimentName, tsrSetType, tsrSet = 1, upstreamDist,
  downstreamDist, featureType, featureColumn, writeTable = TRUE)
}
\arguments{
\item{experimentName}{S4 object of class tssObject with filled data slots tsrData (and/or tsrDataMerged)}

\item{tsrSetType}{Specifies the type of TSR set to be processed.  Options are "replicates" or "merged".}

\item{tsrSet}{Number of the data set of type tsrSetType to be processed}

\item{upstreamDist}{- the maximum distance (in bp) upstream of the selected interval necessary to associate a TSR with a given annotation.}

\item{downstreamDist}{- the maximum distance (in bp) downstream of the selected interval (beginning with the CDS) to associate a TSR with a given annotation.}

\item{featureType}{- Does the annotation file have more than just genes in its 'type' column? Defaults to TRUE.}

\item{featureColumn}{- What is the name of the column in the annotation file containing the featureIDs? Defaults to 'ID'.}

\item{writeTable}{Specifies whether the output should be written to a tab-delimited file. Defaults to TRUE.}
}
\value{
addAnnotationToTSR adds featureID information to the tsrData(Merged) data frame and attaches it to its tssObject slot.
}
\description{
addAnnotationToTSR associates an identified promoter with a given gene, if found upstream and on the same strand within a selected distance.
}

