% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToTSS.R
\docType{methods}
\name{bamToTSS}
\alias{bamToTSS}
\alias{bamToTSS,tssObject-method}
\title{\strong{bamToTSS}}
\usage{
bamToTSS(experimentName)

\S4method{bamToTSS}{tssObject}(experimentName)
}
\arguments{
\item{experimentName}{an S4 object of class tssObject with bam files loaded}
}
\value{
produces a \linkS4class{GRangesList} containing separate
\linkS4class{GRanges} objects for each .bam file contained within
\emph{experimentName}, placing them them in the returned \emph{tssObject}.
}
\description{
\code{bamToTSS} extracts TSS information from each
attached .bam file in a tssObject object
}
\note{
An example similar to the one provided can be
found in the vignette (/inst/doc/TSRchitect.Rmd).
}
\examples{
load(system.file("extdata", "tssObjectExample.RData",
package="TSRchitect"))
tssObjectExample <- bamToTSS(experimentName=tssObjectExample)

}
