% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToTSS.R
\name{bamToTSS}
\alias{bamToTSS}
\title{bamToTSS
Extracts TSS information from each bam file in a tssExp object}
\usage{
bamToTSS(expName)
}
\arguments{
\item{expName}{an object of tssExp format containing bam files}
}
\value{
creates a list of TSSs in class GRanges for each bam file contained within expName.
}
\description{
bamToTSS
Extracts TSS information from each bam file in a tssExp object
}

