% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedToTSS.R
\name{bedToTSS}
\alias{bedToTSS}
\alias{bedToTSS,tssObject-method}
\title{\strong{bedToTSS}}
\usage{
bedToTSS(experimentName)

\S4method{bedToTSS}{tssObject}(experimentName)
}
\arguments{
\item{experimentName}{an S4 object of class tssObject with bed files loaded}
}
\value{
produces a \linkS4class{GRangesList} containing separate
\linkS4class{GRanges} objects for each .bed file contained within
\emph{experimentName}, placing them them in the returned \emph{tssObject}.
}
\description{
\code{bedToTSS} extracts TSS information from each
attached .bed file in a tssObject object
}
\note{
An example similar to the one provided can be
found in the vignette (/inst/doc/TSRchitect.Rmd).
}
