% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cDNAtoBED.R
\docType{methods}
\name{cDNAtoBED}
\alias{cDNAtoBED}
\alias{cDNAtoBED,character,character-method}
\title{\strong{cDNAtoBED}}
\usage{
cDNAtoBED(gsqFile, fileName = "gsqOut.bed")

\S4method{cDNAtoBED}{character,character}(gsqFile,
  fileName = "gsqOut.bed")
}
\arguments{
\item{gsqFile}{a path to the gsq (GeneSeqer) output file (class character)}

\item{fileName}{the name (class character) of the BED file to be written
(default is "gsqOut.bed")}
}
\value{
a BED file containing a list of the 5'-most base from each of the
alignments contained in the GeneSeqer (.gsq) output file is written to the
user's working directory.
}
\description{
\code{cDNAtoBED} converts aligned cDNA data (in .gsq format)
to BED format, extracting the 5'-most base.
}
\examples{
extdata.dir <- system.file("extdata", package="TSRchitect")
tssObjectExample <- cDNAtoBED(gsqFile=paste(extdata.dir,"AtEST.gsq",sep="/"),
                              fileName="testOut.bed")
}
