% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterChr.R
\name{clusterChr}
\alias{clusterChr}
\title{Partitions, then clusters tss data by chromosome.}
\usage{
clusterChr(expName, tssNum, chrName)
}
\arguments{
\item{expName}{an object of class tssExp containing the tss data}

\item{tssNum}{the slot number of the tss data to be clustered}

\item{chrName}{the name of the chromosome to select}
}
\value{
a list object containing xmeans-clustered TSS data for a single chromosome (plus and minus strands)
}
\description{
Partitions, then clusters tss data by chromosome.
}

