% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensusTSRs.R
\name{consensusTSRs}
\alias{consensusTSRs}
\title{Generates a set of consensus TSRs from all conditions within the experiments loaded}
\usage{
consensusTSRs(experimentName)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject that contains information about the experiment}
}
\value{
a data frame of TSRs from all conditions
}
\description{
Generates a set of consensus TSRs from all conditions within the experiments loaded
}

