% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countsMatrix.R
\name{countsMatrix}
\alias{countsMatrix}
\title{Creates a data frame with a list of TSS positions and the corresponding counts from each}
\usage{
countsMatrix(experimentName)
}
\arguments{
\item{experimentName}{an S4 object of class tssObject that contains information about the experiment (including tssData)}
}
\value{
Assigns a data frame to the expData slot of your tssObject
}
\description{
Creates a data frame with a list of TSS positions and the corresponding counts from each
}

