% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr-metrics.R
\name{countsToVector}
\alias{countsToVector}
\title{countsToVector}
\usage{
countsToVector(x)
}
\value{
a data frame containing the coordinates tag counts for each
unique TSS.
}
\description{
An internal function that is utilized during tsrCluster.
countsToVector is invoked by the internal function clusterTSR during
the TSR identification process.
}
\keyword{internal}
