% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSummarizedExperiment.R
\name{createSummarizedExperiment}
\alias{createSummarizedExperiment}
\alias{createSummarizedExperiment,tssObject,character,numeric,character-method}
\title{\strong{createSummarizedExperiment}}
\usage{
createSummarizedExperiment(
  experimentName,
  tsrSetType = "merged",
  tsrSet = 1,
  samplePrefix
)

\S4method{createSummarizedExperiment}{tssObject,character,numeric,character}(
  experimentName,
  tsrSetType = "merged",
  tsrSet = 1,
  samplePrefix
)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject} containing
information in slot \emph{@tssTagData}}

\item{tsrSetType}{specifies the TSR set to be converted into a
\emph{SummarizedExperiment} object. Options are "replicates" or "merged".
(character)}

\item{tsrSet}{number of the dataset to be processed (numeric).}

\item{samplePrefix}{the prefix (or prefixes) that match the sample
identifiers in the tsrData column. (character)}
}
\value{
a summarizedExperiment object from the specified TSR data set
that is to be written to your working directory.
}
\description{
\code{createSummarizedExperiment} creates and reutrns
a \emph{SummarizedExperiment} object from the tag counts on a selected 
TSR dataset.
}
\note{
For more information on the SummarizedExperiment class, please visit
https://bioconductor.org/packages/release/bioc/vignettes/SummarizedExperiment/inst/doc/SummarizedExperiment.html
}
\examples{
load(system.file("extdata", "tssObjectExample.RData", package="TSRchitect"))
createSummarizedExperiment(tssObjectExample, tsrSetType="merged", tsrSet=1,
samplePrefix=c("sample1","sample2"))

}
