% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineTSR.R
\name{determineTSR}
\alias{determineTSR}
\title{\emph{determineTSR}}
\usage{
determineTSR(experimentName, parallel, tsrSetType, tssSet, tagCountThreshold,
  clustDist, writeTable = FALSE)
}
\arguments{
\item{experimentName}{an object of class \emph{tssObject} containing information in slot \emph{@tssTagData}}

\item{parallel}{if TRUE, the analysis is run in parallel (logical)}

\item{tsrSetType}{specifies the set to be clustered. Options are "replicates" or "merged". (character)}

\item{tssSet}{default is "all"; if a single TSS dataset is desired, specify tssSet number (character)}

\item{tagCountThreshold}{the number of TSSs required at a given position for it to be considered in TSR identification. (numeric)}

\item{clustDist}{the maximum distance of TSSs between two TSRs in base pairs. (numeric)}

\item{writeTable}{specifies whether the output should be written to a table. (logical)}
}
\value{
creates a list of \linkS4class{GenomicRanges}-containing TSR positions in slot \emph{@tsrData} on the \emph{tssObject} object
}
\description{
\code{determineTSR} Identifies TSRs from entire TSS datasets as specified
}
\note{
The example provided can be found in \emph{Example 1} from the vignette (/inst/doc/TSRchitect.Rmd)
}
\examples{
determineTSR(experimentName = Hs_RAMPAGE, parallel = TRUE, tsrSetType = "replicates", tssSet="all", tagCountThreshold=25, clustDist=20, writeTable=TRUE)
}

