% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detTSR.R
\name{detTSR}
\alias{detTSR}
\title{detTSR
Finds TSRs from a given sequence (Internal function)}
\usage{
detTSR(experimentName, tsrSetType, tssSet = 1, tagCountThreshold, clustDist)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject containing
information in slot tssTagData}

\item{tsrSetType}{- specifies the set to be clustered.
Options are "replicates" or "merged".}

\item{tssSet}{- number of the dataset to be analyzed}

\item{tagCountThreshold}{- number of TSSs required at a given position}

\item{clustDist}{- maximum distance of TSSs between two TSRs (in base pairs)}
}
\value{
creates a list of GenomicRanges containing TSR positions in slot
'tsrData' on your tssObject object
}
\description{
detTSR
Finds TSRs from a given sequence (Internal function)
}
\keyword{internal}
