% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{expressionCTSS}
\alias{expressionCTSS}
\title{expressionCTSS
Returns a matrix [a, h] where a = the number of unique TSSs and h = the # of tags observed at that position}
\usage{
expressionCTSS(x)
}
\description{
expressionCTSS
Returns a matrix [a, h] where a = the number of unique TSSs and h = the # of tags observed at that position
}

