% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{expressionTSS}
\alias{expressionTSS}
\title{expressionTSS
Returns a matrix [a, h] where a = the number of unique TSSs and h = the # of tags observed at that position}
\usage{
expressionTSS(x, dfName = "TSS.txt", writeDF = FALSE)
}
\description{
expressionTSS
Returns a matrix [a, h] where a = the number of unique TSSs and h = the # of tags observed at that position
}

