% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R
\docType{methods}
\name{getBamDataFirstRead}
\alias{getBamDataFirstRead}
\alias{getBamDataFirstRead,tssObject,numeric-method}
\title{\strong{getBamDataFirstRead}}
\usage{
getBamDataFirstRead(experimentName, slot)

\S4method{getBamDataFirstRead}{tssObject,numeric}(experimentName, slot)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject}}

\item{slot}{'numeric' a number corresponding to the slot in
"bamDataFirstRead" to be retrieved.}
}
\value{
the contents of the specified slot "bamDataFirstRead" are returned
}
\description{
an accessor function that retrieves the contents of
a specified slot "bamDataFirstRead" from a given \emph{tssObject}
}
\examples{
load(system.file("extdata", "tssObjectExample.RData",
package="TSRchitect"))
example.bamDataFirstRead <-
getBamDataFirstRead(experimentName=tssObjectExample, slot = 1)
example.bamDataFirstRead

}
\keyword{methods}
