% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R
\docType{methods}
\name{getBamDataLastRead}
\alias{getBamDataLastRead}
\alias{getBamDataLastRead,tssObject,numeric-method}
\title{\strong{getBamDataLastRead}}
\usage{
getBamDataLastRead(experimentName, slot)

\S4method{getBamDataLastRead}{tssObject,numeric}(experimentName, slot)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject}}

\item{slot}{'numeric' a number corresponding to the slot in
"bamDataLastRead" to be retrieved.}
}
\value{
the contents of the specified slot "bamDataLastRead" are returned
}
\description{
an accessor function that retrieves the contents of
a specified slot "bamDataLastRead" from a given \emph{tssObject}
}
\keyword{methods}
