% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R
\name{getFileNames}
\alias{getFileNames}
\title{\strong{getFileNames}}
\usage{
getFileNames(experimentName)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject}}
}
\value{
the contents of slot "fileNames" are returned, which are
a vector of class "character"
}
\description{
an accessor function that retrieves the contents of slot
"fileNames" from a given \emph{tssObject}
}
\examples{
load(system.file("extdata", "tssObjectExample.RData",
package="TSRchitect"))
example.file.names <- getFileNames(experimentName=tssObjectExample)
example.file.names

}
\keyword{methods}
