% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPathBed.R
\docType{methods}
\name{getPathBed}
\alias{getPathBed}
\alias{getPathBed,tssObject,numeric,logical-method}
\title{getPathBed}
\usage{
getPathBed(experimentName, setNum, is.bedpe)

\S4method{getPathBed}{tssObject,numeric,logical}(experimentName, setNum,
  is.bedpe = FALSE)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject containing
information in slot tssTagData}

\item{setNum}{- a numeric value representing the fileName slot
to access}
}
\value{
returns a single GRanges object from a BED file (internal)
}
\description{
An internal function  which is invoked using the user-level
function loadTSSObject to import BED data.
(Internal function)
}
