% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R
\name{getTitle}
\alias{getTitle}
\title{\strong{getTitle}}
\usage{
getTitle(experimentName)
}
\arguments{
\item{experimentName}{n S4 object of class \emph{tssObject}}
}
\value{
the contents of slot "title" are returned
}
\description{
an accessor function that retrieves the contents of slot "title"
from a given \emph{tssObject}
}
\examples{
load(system.file("extdata", "tssObjectExample.RData",
package="TSRchitect"))
example.title <- getTitle(experimentName=tssObjectExample)
example.title

}
\keyword{methods}

