% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R
\name{getTSRdata}
\alias{getTSRdata}
\alias{getTSRdata,tssObject,character,numeric-method}
\title{\strong{getTSRdata}}
\usage{
getTSRdata(experimentName, slotType, slot)

\S4method{getTSRdata}{tssObject,character,numeric}(experimentName, slotType = c("replicates", "merged"), slot)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject}}

\item{slotType}{'character' which data type is to be selected.
Either "replicates" (tsrCountData) or "merged" (tsrCountDataMerged)}

\item{slot}{'numeric' a number corresponding to the slot in
"tsrData"/"tsrDataMerged" to be retrieved.}
}
\value{
the contents of the selected slot (either "tsrData" or
"tsrDataMerged" are returned)
}
\description{
an accessor function that retrieves the contents of
a specified slot "tsrData"/"tsrDataMerged" from a
given \emph{tssObject}
}
\examples{
load(system.file("extdata", "tssObjectExample.RData",
package="TSRchitect"))
ex.tsrData <- getTSRdata(experimentName=tssObjectExample,
slotType="replicates", slot = 1)
ex.tsrData

}
\keyword{methods}
