% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R
\name{getTSScountData}
\alias{getTSScountData}
\title{\strong{getTSScountData}}
\usage{
getTSScountData(experimentName, slotType, slot)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject}}

\item{slotType}{'character' which data type is to be selected.
Either "replicates" (tssCountData) or "merged" (tssCountDataMerged)}

\item{slot}{'numeric' a number corresponding to the slot in
"tssCountData"/"tssCountDataMerged" to be retrieved.}
}
\value{
the contents of the selected slot (either "tssCountData" or
"tssCountDataMerged" are returned)
}
\description{
an accessor function that retrieves the contents of
a specified slot "tssCountData"/"tssCountDataMerged" from a
given \emph{tssObject}
}
\examples{
load(system.file("extdata", "tssObjectExample.RData",
package="TSRchitect"))
ex.tssCountData <- getTSScountData(experimentName=tssObjectExample,
slotType="replicates", slot = 1)
ex.tssCountData

}
\keyword{methods}
