% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get-methods.R
\name{getTSStagData}
\alias{getTSStagData}
\title{\strong{getTSStagData}}
\usage{
getTSStagData(experimentName, slot)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject}}

\item{slot}{'numeric' a number corresponding to the slot in
"tssTagData" to be retrieved.}
}
\value{
the contents of the specified slot "tssTagData" are returned
}
\description{
an accessor function that retrieves the contents of
a specified slot "tssTagData" from a given \emph{tssObject}
}
\examples{
load(system.file("extdata", "tssObjectExample.RData",
package="TSRchitect"))
example.tssTagData <- getTSStagData(experimentName=tssObjectExample, slot=1)
example.tssTagData

}
\keyword{methods}

