% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAnnotation.R
\name{importAnnotation}
\alias{importAnnotation}
\title{\strong{importAnnotation}}
\usage{
importAnnotation(experimentName, fileType, annotFile)
}
\arguments{
\item{experimentName}{- an S4 object of class \emph{tssObject} that contains information about the experiment}

\item{fileType}{- the format of the annotation file to be imported. Must be one of: "bed", "gff" or "gff3".}

\item{annotFile}{- a path (full or relative) to the annotation file to be imported.}
}
\value{
\emph{importAnnotation} fills the slot experimentName@annotation in the tssObject \emph{experimentName} with
        a \linkS4class{GRanges} object contining the annotation data.
}
\description{
\code{importAnnotation} imports an annotation from an external file and attaches it to \emph{experimentName}.
}
\note{
The example provided can be found in \emph{Example 1} from the vignette (/inst/doc/TSRchitect.Rmd)

\code{importAnnotation} makes use of three functions from the \emph{rtracklayer} Bioconductor package: \code{\link[rtracklayer]{import.bed}}, \code{\link[rtracklayer]{import.gff}}, and \code{\link[rtracklayer]{import.gff3}}
}
\examples{
importAnnotation(experimentName = Hs_RAMPAGE, fileType ="gff3", annotFile ="vignettes/gencode.v19.annotation.gff3")
}

