% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAnnotationHub.R
\docType{methods}
\name{importAnnotationHub}
\alias{importAnnotationHub}
\alias{importAnnotationHub,tssObject,character,character,character,character-method}
\title{\strong{importAnnotationHub}}
\usage{
importAnnotationHub(experimentName, provider, annotType, species, annotID)


  \S4method{importAnnotationHub}{tssObject,character,character,character,character}(experimentName,
  provider, annotType, species, annotID)
}
\arguments{
\item{experimentName}{- an S4 object of class \emph{tssObject}
that contains information about the experiment}

\item{provider}{- 'character' the source of the annotation in AnnotationHub
(e.g. 'gencode')}

\item{annotType}{- 'character' the format of the annotationHub annotation
to be imported. Using 'gff' will find annotations in both gff or gff3 formats}

\item{species}{- 'character' the species identifier in AnnotationHub
(e.g. 'human')}

\item{annotID}{- 'character' the AnnotationHub identifier to be retrieved}
}
\value{
fills the slot \emph{@annotation} in the returned \emph{tssObject}
with an AnnotationHub record. The record retrieved must be an object
of class \linkS4class{GRanges}.
}
\description{
imports an annotation directly from AnnotationHub
and attaches it to the \emph{tssObject}
}
\note{
An example similar to the one provided can be found in
the vignette (/inst/doc/TSRchitect.Rmd)

Please consult the available records in AnnotationHub
beforehand using \code{\link[AnnotationHub]{AnnotationHub}}
}
