% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBam.R
\name{importBam}
\alias{importBam}
\title{Computes TSS positions from all bam files and loads them into a tssObject object}
\usage{
importBam(experimentName)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject that contains information about the experiment}
}
\value{
alignment data (in BAM format) from the tss profiling experiments assigned to your tssObject object
}
\description{
Computes TSS positions from all bam files and loads them into a tssObject object
}

