% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTSSobj.R
\name{loadTSSobj}
\alias{loadTSSobj}
\title{\strong{loadTSSobj}}
\usage{
loadTSSobj(experimentTitle, inputDir, inputType, isPairedEnd, sampleNames,
  replicateIDs)
}
\arguments{
\item{inputDir}{path to the directory containing the alignment files in .bam
format (character).
Note that all the paths to all files in \emph{inputDir} with the extension
.bam in \emph{inputDir} will be imported with this function.}

\item{experimentName}{an S4 object of class \emph{tssObject} that
contains information about the experiment}
}
\value{
\emph{loadTSSobj} fills the slot \emph{bamData} on the returned
\emph{tssObject} with \linkS4class{GAlignments} objects from the
\bold{GenomicAlignments} package, one for each attached .bam file
on the \emph{fileNames} slot.
}
\description{
\code{loadTSSobj} processes alignment files in .bam format
from the local directory supplied.
}
\note{
An example similar to the one provided can be found in
the vignette (/inst/doc/TSRchitect.Rmd).

All .bam files found in \emph{inputDir} will be
retrieved and written in ascending alphanumeric order to the
\emph{@fileNames} slot on the \emph{tssObject} that is created.
}
\examples{
load(system.file("extdata", "tssObjectExample.RData", package="TSRchitect"))
extdata.dir <- system.file("extdata", package="TSRchitect")
loadTSSobj(experimentName=tssObjectExample, inputDir=extdata.dir)

}
