% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGRangesFromTSR.R
\name{makeGRangesFromTSR}
\alias{makeGRangesFromTSR}
\title{\strong{makeGRangesFromTSR}}
\usage{
makeGRangesFromTSR(experimentName, tsrSetType, tsrSet = 1)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject} containing
information in slot \emph{@tssTagData}}

\item{tsrSetType}{specifies the set to be written to file.
Options are "replicates" or "merged". (character)}

\item{tsrSet}{number of the dataset to be processed (numeric).}
}
\value{
An object of class \emph{GRanges} containing the specified TSR
data set. Headers include 'seqnames', 'ranges' (including start and end),
'strand', 'name' (TSR ID) and 'score' (Shape Index/SI) value
}
\description{
\code{makeGRangesFromTSR} creates a GRanges object
from a specified TSR data set
}
\note{
For more information on the GRanges class, please visit:
http://web.mit.edu/~r/current/arch/i386_linux26/lib/R/library/GenomicRanges/html/GRanges-class.html
}
\examples{
load(system.file("extdata", "tssObjectExample.RData", package="TSRchitect"))
makeGRangesFromTSR(experimentName=tssObjectExample, tsrSetType="replicates",
tsrSet=1)

}
