% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSampleData.R
\name{mergeSampleData}
\alias{mergeSampleData}
\alias{mergeSampleData,tssObject,numeric,numeric-method}
\title{\strong{mergeSampleData}}
\usage{
mergeSampleData(experimentName, n.cores, tagCountThreshold)

\S4method{mergeSampleData}{tssObject,numeric,numeric}(experimentName, n.cores = 1, tagCountThreshold = 1)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject} that contains
information about the experiment.}

\item{n.cores}{the number of cores to be used for this job.
ncores=1 means serial execution of function calls (numeric)}

\item{tagCountThreshold}{the number of TSSs required at a given position
for it to be considered in sample data merging. (numeric)
Note: Merged data sets can become very large when the tagCountThreshold is
      set low (leading to inclusion of a lot of "noise" and resulting in
      long execution times in the necessary TSS position ordering step).}
}
\value{
tssCountData datasets are merged (according to the
\emph{sampleIDs}) and put in the tssCountDataMerged slot in the returned
\emph{tssObject}.
}
\description{
\code{mergeSampleData} combines samples from multiple TSS
experiments into a single \linkS4class{GRanges} object
}
\note{
An example similar to the one provided can be found in
the vignette (/inst/doc/TSRchitect.Rmd).
}
\examples{
load(system.file("extdata", "tssObjectExample.RData",
package="TSRchitect"))
tssObjectExample <- mergeSampleData(experimentName=tssObjectExample,
                                    n.cores=1, tagCountThreshold=1)

}
