% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSampleData.R
\name{mergeSampleData}
\alias{mergeSampleData}
\title{\strong{mergeSampleData}}
\usage{
mergeSampleData(experimentName)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject} that contains information about the experiment.}
}
\value{
tssCountData datasets will be merged (according to the \emph{sampleIDs}) and assigned to your \emph{tssObject}.
}
\description{
\code{mergeSampleData} combines samples from two different TSS experiments into a single \linkS4class{GRanges} object
}
\note{
The example provided can be found in \emph{Example 1} from the vignette (/inst/doc/TSRchitect.Rmd).
}
\examples{
mergeSampleData(experimentName=Hs_RAMPAGE)
}

