% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTSS.R
\name{mergeTSS}
\alias{mergeTSS}
\title{Combines samples from different tss experiments into a single GRanges object}
\usage{
mergeTSS(expName)
}
\arguments{
\item{expName}{- an S4 object of class tssExp that contains information about the experiment.}
}
\value{
merged tss profiling experiments according to their assigned sampleIDs to your tssExp object
}
\description{
Combines samples from different tss experiments into a single GRanges object
}

