% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTSSp.R
\name{processTSSp}
\alias{processTSSp}
\title{processTSSp
Creates an expression matrix for all TSSs within a given TSS experiment (in tssTagData)}
\usage{
processTSSp(experimentName, tssSet, writeTable)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject containing information in slot tssTagData}

\item{tssSet}{- number of the dataset to be analyzed}

\item{writeTable}{if TRUE, writes a data frame containing the TSSs positions and their abundance to your workspace}
}
\value{
creates a data frame containing tss expression for each TSS to be stored in slot 'tssExpression' on your tssObject object
}
\description{
processTSSp
Creates an expression matrix for all TSSs within a given TSS experiment (in tssTagData)
}

