% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setSampleID.R
\name{setSampleID}
\alias{setSampleID}
\title{\strong{bamToTSS}}
\usage{
setSampleID(experimentName, sample.names, replicate.IDs)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject} that
contains information about the experiment.}

\item{sample.names}{unique labels of class character for each TSS sample
within the experiment.}

\item{replicate.IDs}{identifiers indicating which samples are biological
replicates. As with \emph{sample.names}, note that \code{importBam} imports
bam data in ascending alphanumeric order, so replicate.IDs must be arranged
in this order also so that they directly correspond to the intended file.}
}
\value{
names and replicate information for experimental samples added
to the returned \emph{tssObject} object.
}
\description{
\code{bamToTSS} Sets names and replicate information
for experimental samples in a \emph{tssObject} S4 object.
}
\note{
Please note that \code{importBam} attaches .bam data in ascending
alphanumeric order, so \emph{sample.names} must be arranged in such
a way that they directly correspond to the intended file.
This order can be specified using \code{\link{getFileNames}}
}
\examples{
load(system.file("extdata", "tssObjectExample.RData",
package="TSRchitect"))
setSampleID(experimentName=tssObjectExample,
sample.names=c("sample1-rep1","sample1-rep2","sample2-rep1","sample2-rep2"),
replicate.IDs=c(1,1,2,2))
#experiments 1 & 2 and 3 & 4 are replicates of samples 1 and 2, respectively

}

