% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeLibrary.R
\name{summarizeLibrary}
\alias{summarizeLibrary}
\title{Summarizes the data for a TSS profiling experiment and loads them to an object of class tssObjet}
\usage{
summarizeLibrary(experimentName)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject that contains information about the experiment}
}
\value{
a tabular summary of the tss profiling experiments assigned to your tssObject
}
\description{
Summarizes the data for a TSS profiling experiment and loads them to an object of class tssObjet
}

