% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tagCountTSS}
\alias{tagCountTSS}
\title{tagCountTSS
Returns a matrix [a, h] where a = the number of unique TSSs and h = the # of tags observed at that position}
\usage{
tagCountTSS(x, outfname = "TSS.txt", writeDF = FALSE)
}
\description{
tagCountTSS
Returns a matrix [a, h] where a = the number of unique TSSs and h = the # of tags observed at that position
}

