% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSRchitectUsersGuide.R
\name{TSRchitectUsersGuide}
\alias{TSRchitectUsersGuide}
\title{\strong{TSRchitectUsersGuide}}
\usage{
TSRchitectUsersGuide(view = TRUE)
}
\value{
Opens the TSRchitect User's Guide in a pdf viewer on the user's system.
}

