% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tsrCluster}
\alias{tsrCluster}
\title{tsrCluster
Partitions, then clusters tss data by sequence. (Internal function)
returns a list of TSRs from the data.frame generated by tagCountTSS()}
\usage{
tsrCluster(x, minNbrTSSs = 3, minDist = 20)
}
\description{
tsrCluster
Partitions, then clusters tss data by sequence. (Internal function)
returns a list of TSRs from the data.frame generated by tagCountTSS()
}

