% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr-metrics.R, R/tsrCounts.R
\name{tsrCounts}
\alias{tsrCounts}
\title{An internal function that caculates the total number of counts associated with a given TSR}
\usage{
tsrCounts(expName)

tsrCounts(expName)
}
\arguments{
\item{expName}{- a S4 object of class tssExp that contains information about the experiment}
}
\value{
alignment data (in BAM format) from the tss profiling experiments assigned to your tssExp object
}
\description{
An internal function that caculates the total number of counts associated with a given TSR

Creates a data frame that contains the expression of each
}

