% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrFind.R
\name{tsrFind}
\alias{tsrFind}
\title{tsrFind
Finds TSRs from a given chromosome}
\usage{
tsrFind(expName, tssNum = 1, nTSSs, clustDist, setToCluster,
  writeTable = FALSE)
}
\arguments{
\item{expName}{- a S4 object of class tssExp containing information in slot tssData}

\item{tssNum}{- number of the dataset to be analyzed}

\item{nTSSs}{- number of TSSs required at a given position}

\item{clustDist}{- maximum distance of TSSs between two TSRs (in base pairs)}

\item{setToCluster}{- specifies the set to be clustered. Options are "replicates" or "merged".}

\item{writeTable}{- specifies whether the output should be written to a table. (logical)}
}
\value{
creates a list of GenomicRanges containing TSR positions in slot 'tsrData' on your tssExp object
}
\description{
tsrFind
Finds TSRs from a given chromosome
}

