% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{TSRmshapeIndex}
\alias{TSRmshapeIndex}
\title{TSRmshapeIndex}
\usage{
TSRmshapeIndex(tssArray)
}
\arguments{
\item{tssArray}{an object containing TSS coordinates and their
abundances. (data.frame)}
}
\value{
Returns a  numeric vector of length 1 containing the Modified Shape Index (mSI)
value for the selected TSR.
}
\description{
An internal function that caculates the modified shape index (mSI)
of a given TSR from the output of tsrCluster.
TSRmshapeIndex is called by tssArrayProperties().
}
\keyword{internal}
