% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrToAnnotation.R
\name{tsrToAnnotation}
\alias{tsrToAnnotation}
\title{tsrToAnnotation}
\usage{
tsrToAnnotation(experimentName, upstreamDist, downstreamDist, featureType,
  featureColumn)
}
\arguments{
\item{experimentName}{- an S4 object of class tssObject that contains information about the experiment.}

\item{upstreamDist}{- the maximum distance (in bp) upstream of the selected interval necessary to associate a TSR with a given annotation.}

\item{downstreamDist}{- the maximum distance (in bp) downstream of the selected interval (beginning with the CDS) to associate a TSR with a given annotation.}

\item{featureType}{- Does the annotation file have more than just genes in its 'type' column? Defaults to TRUE.}

\item{featureColumn}{- What is the name of the column in the annotation file containing the featureIDs? Defaults to 'ID'.}
}
\value{
tsrToAnnotation adds featureID information to the tsrData data frame and attaches it to the tssObject
experimentName
}
\description{
tsrToAnnotation associates an identified promoter with a given gene, if found upstream and on the same strand within a selected distance.
}

