% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrToBed.R
\name{tsrToBed}
\alias{tsrToBed}
\title{Writes the contents of a slot in TSRdata to a BED file}
\usage{
tsrToBed(expName, Name, writeBed)
}
\arguments{
\item{expName}{an object of class tssExp with TSR data loaded}

\item{Name}{the name of the BED file to be written in your working directory}
}
\description{
Writes the contents of a slot in TSRdata to a BED file
}

