% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrToDF.R
\name{tsrToDF}
\alias{tsrToDF}
\title{An internal function for TSS clustering
Creates a data frame for a given output of .tsrCluster}
\usage{
tsrToDF(x)
}
\description{
An internal function for TSS clustering
Creates a data frame for a given output of .tsrCluster
}

