% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr-metrics.R
\name{tsrWidth}
\alias{tsrWidth}
\title{An internal function that caculates the width of a given TSR from the output of tsrCluster}
\usage{
tsrWidth(x)
}
\description{
An internal function that caculates the width of a given TSR from the output of tsrCluster
}

