% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tssChr}
\alias{tssChr}
\title{tssChr}
\usage{
tssChr(tssObj, seqName)
}
\arguments{
\item{tssObj}{an object of class \linkS4class{GRanges} containing data
from a slot of \emph{tssTagData}}

\item{seqName}{the name of the sequence to select}
}
\value{
generates a list containing TSS data for a single sequence
(plus and minus strands)
}
\description{
Retrieves tss data from a given experiment by sequence.
}
\keyword{internal}
