% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssExpr.R
\name{tssExpr}
\alias{tssExpr}
\title{tssExpr
Creates an expression matrix for all TSSs within a given TSS experiment (in tssData)}
\usage{
tssExpr(expName, tssNum, writeTable)
}
\arguments{
\item{expName}{an object of tssExp format containing information in slot tssData}

\item{tssNum}{the number of the dataset to be analyzed}

\item{writeTable}{if TRUE, writes a data frame containing the TSSs positions and their abundance to your workspace}
}
\value{
creates a data frame containing tss expression for each CTSS in slot 'tssExpr' on your tssExp object
}
\description{
tssExpr
Creates an expression matrix for all TSSs within a given TSS experiment (in tssData)
}

