% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssExprP.R
\name{tssExprP}
\alias{tssExprP}
\title{tssExprP
Creates an expression matrix for all TSSs within a given TSS experiment (in tssData)}
\usage{
tssExprP(expName, tssNum, writeTable)
}
\arguments{
\item{expName}{- a S4 object of class tssExp containing information in slot tssData}

\item{tssNum}{- number of the dataset to be analyzed}

\item{writeTable}{if TRUE, writes a data frame containing the TSSs positions and their abundance to your workspace}
}
\value{
creates a data frame containing tss expression for each CTSS in slot 'tssExpr' on your tssExp object
}
\description{
tssExprP
Creates an expression matrix for all TSSs within a given TSS experiment (in tssData)
}

