% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Allclasses.R
\name{tssObject}
\alias{tssObject}
\title{\strong{tssObject}}
\usage{
tssObject(title = NA, bamData = NA)
}
\arguments{
\item{title}{'character' A short descriptive title for the
experiment. Is set to NA by default.}

\item{bamData}{'list' the name of a list of \linkS4class{GAlignments}
objects in the workspace. Set to NA by default.}
}
\value{
a new \emph{tssObject} is returned to the user's workspace.
}
\description{
S4 constructor function for \emph{tssObject}
}
\examples{
new.tssObj <- tssObject(title="Example")

}
