% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssToBed.R
\name{tssToBed}
\alias{tssToBed}
\title{Converts TSS data into bed format}
\usage{
tssToBed(expName, fileName = "myTSS", allData = TRUE)
}
\arguments{
\item{TSS}{data from a tssExp object}
}
\description{
Converts TSS data into bed format
}

