% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTSR.R
\name{writeTSR}
\alias{writeTSR}
\title{writeTSR
Write TSRs for a specified data set to a tab-delimited file}
\usage{
writeTSR(experimentName, tsrSetType, tsrSet = 1, fileType = "tab")
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject containing information in slot tssTagData}

\item{tsrSetType}{- specifies the set to be written to file. Options are "replicates" or "merged".}

\item{tsrSet}{- number of the dataset to be processed}

\item{fileType}{- "tab" for tab-delimited output,"bed" for bed-file}
}
\value{
a table containing the specified TSR data set will be written to your working directory
}
\description{
writeTSR
Write TSRs for a specified data set to a tab-delimited file
}

