% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTSS.R
\docType{methods}
\name{writeTSS}
\alias{writeTSS}
\alias{writeTSS,tssObject,character,numeric,character,character-method}
\title{\strong{writeTSS}}
\usage{
writeTSS(experimentName, tssSetType, tssSet = 1, tssLabel = "TSS_",
  fileType = "tab")


  \S4method{writeTSS}{tssObject,character,numeric,character,character}(experimentName,
  tssSetType, tssSet = 1, tssLabel = "TSS_", fileType = "tab")
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject} containing
information in slot \emph{@tssCountData}}

\item{tssSetType}{specifies the set to be written to file.
Options are "replicates" or "merged". (character)}

\item{tssSet}{number of the dataset to be processed (numeric).}

\item{tssLabel}{specifies the label to be used in the name column of
BED format output}

\item{fileType}{the format of the file to be written.
Possible choices are "tab" for tab-delimited output, "bed" for BED format,
and "bedGraph" for for GFF3 format (character).}
}
\value{
A table containing the specified TSS data set that
is to be written to your working directory.
}
\description{
\code{writeTSS} writes identified TSSs
from a specified data set to a file in either tab or BED formats
}
\note{
The .bed file written adheres to the standard six-column BED format,
while "tab" format is identical to that of the data.frames containing TSS
data.

For more information on the BED format, please visit
https://genome.ucsc.edu/FAQ/FAQformat#format1
}
\examples{
load(system.file("extdata", "tssObjectExample.RData", package="TSRchitect"))
writeTSS(experimentName=tssObjectExample, tssSetType="replicates",
         tssSet=1, tssLabel="TSSsample1_",fileType="tab")

}
