% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{acquireTSS}
\alias{acquireTSS}
\title{acquireTSS}
\usage{
acquireTSS(experimentName, tssSet)
}
\arguments{
\item{experimentName}{an S4 object of class tssObject
containing the tss data}

\item{tssSet}{the slot number of the tss data to be clustered}
}
\value{
a list object containing TSS data from the entire dataset
}
\description{
Retrieves all tss data from a given TSS experiment.
}
\keyword{internal}
