% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAnnotationToTSR.R
\name{addAnnotationToTSR}
\alias{addAnnotationToTSR}
\title{addAnnotationToTSR}
\usage{
addAnnotationToTSR(experimentName, tsrSetType, tsrSet = 1, upstreamDist,
  downstreamDist, feature, featureColumnID, writeTable = TRUE)
}
\arguments{
\item{experimentName}{S4 object of class tssObject with filled data slots tsrData (and/or tsrDataMerged)}

\item{tsrSetType}{Specifies the type of TSR set to be processed.  Options are "replicates" or "merged".}

\item{tsrSet}{Number of the data set of type tsrSetType to be processed}

\item{upstreamDist}{- the maximum distance (in bp) upstream of the selected interval necessary to associate a TSR with a given annotation.}

\item{downstreamDist}{- the maximum distance (in bp) downstream of the selected interval (beginning with the CDS) to associate a TSR with a given annotation.}

\item{feature}{Specifies the feature to be used for annotation (typically "gene" [default] or "mRNA" for GFF3 input); set to "all" if all annotations from the input are to be used.}

\item{featureColumnID}{Name of the column identifier in the annotation GRanges object. This should be "ID" (default) for GFF3 input or "name" for bed input.}

\item{writeTable}{Specifies whether the output should be written to a tab-delimited file. Defaults to TRUE.}
}
\value{
addAnnotationToTSR adds feature annotation to the tsrData(Merged) data frame and attaches it to its tssObject slot.
}
\description{
addAnnotationToTSR associates an identified promoter with a given gene, if found upstream and on the same strand within a selected distance.
}

