% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addTagCountsToTSR.R
\name{addTagCountsToTSR}
\alias{addTagCountsToTSR}
\title{addTagCountsToTSR
Adds a column of tag counts to a set of identifid TSRs}
\usage{
addTagCountsToTSR(experimentName, tsrSetType, tsrSet = 1,
  tagCountThreshold = 1, writeTable = TRUE)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject containing information in slot tssTagData}

\item{tsrSetType}{- specifies the set to be written to file. Options are "replicates" or "merged".}

\item{tsrSet}{- number of the dataset to be processed}

\item{tagCountThreshold}{- number of TSSs required at a given position}

\item{writeTable}{- specifies whether the output should be written to a table. (logical)}
}
\value{
a column of tag counts is appended to the selected set of identified TSRs
}
\description{
addTagCountsToTSR
Adds a column of tag counts to a set of identifid TSRs
}

