% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamToTSS.R
\name{bamToTSS}
\alias{bamToTSS}
\title{bamToTSS
Extracts TSS information from each bam file in a tssObject object}
\usage{
bamToTSS(experimentName)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject with bam files loaded}
}
\value{
creates a list of TSSs in class GRanges for each bam file contained within experimentName and places them in the tssObject object
}
\description{
bamToTSS
Extracts TSS information from each bam file in a tssObject object
}

