% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionMatrix.R
\name{countsMatrix}
\alias{countsMatrix}
\title{Creates a data frame with a list of TSS positions and the corresponding counts from each}
\usage{
countsMatrix(expName)
}
\arguments{
\item{expName}{an S4 object of class tssExp that contains information about the experiment (including tssData)}
}
\value{
Assigns a data frame to the expData slot of your tssExp object
}
\description{
Creates a data frame with a list of TSS positions and the corresponding counts from each
}

