% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr-metrics.R
\name{countsToVector}
\alias{countsToVector}
\title{countsToVector}
\usage{
countsToVector(x)
}
\description{
An internal function that extracts the values
from the output returned from tsrCluster
}
\keyword{internal}

