% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determineTSR.R
\name{determineTSR}
\alias{determineTSR}
\title{determineTSR
Finds TSRs from a given sequence}
\usage{
determineTSR(experimentName, setToCluster, tagCountThreshold, clustDist,
  writeTable = FALSE)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject containing information in slot tssTagData}

\item{setToCluster}{- specifies the set to be clustered. Options are "replicates" or "merged".}

\item{tagCountThreshold}{- number of TSSs required at a given position}

\item{clustDist}{- maximum distance of TSSs between two TSRs (in base pairs)}

\item{writeTable}{- specifies whether the output should be written to a table. (logical)}
}
\value{
creates a list of GenomicRanges containing TSR positions in slot 'tsrData' on your tssObject object
}
\description{
determineTSR
Finds TSRs from a given sequence
}

