% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAnnotation.R
\name{importAnnotation}
\alias{importAnnotation}
\title{\emph{importAnnotation}}
\usage{
importAnnotation(experimentName, fileType, annotFile)
}
\arguments{
\item{experimentName}{- an S4 object of class tssObject that contains information about the experiment}

\item{fileType}{- the format of the annotation file to be imported. Must be one of: "bed", "gff" or "gff3".}

\item{annotFile}{- a path (full or relative) to the annotation file to be imported.}
}
\value{
\emph{importAnnotation} fills the slot experimentName@geneAnnot in the tssObject \emph{experimentName} with
        a GRanges object contining the annotation data
}
\description{
\emph{importAnnotation} imports an annotation from an external file and attaches it to your tssObject.
}

