% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBam.R
\name{importBam}
\alias{importBam}
\title{Computes TSS positions from all from a tssExp object}
\usage{
importBam(expName)
}
\arguments{
\item{expName}{an S4 object of class tssExp that contains information about the experiment.}
}
\value{
bam data from the tss profiling experiments assigned to your tssExp object (expName)
}
\description{
Computes TSS positions from all from a tssExp object
}

