% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBam.R
\name{importBam}
\alias{importBam}
\title{Computes TSS positions from all bam files and loads them into a tssExp object}
\usage{
importBam(expName)
}
\arguments{
\item{expName}{- a S4 object of class tssExp that contains information about the experiment}
}
\value{
bam data from the tss profiling experiments assigned to your tssExp object
}
\description{
Computes TSS positions from all bam files and loads them into a tssExp object
}

