% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importBam.R
\name{importBam}
\alias{importBam}
\title{\emph{importBam()}}
\usage{
importBam(experimentName)
}
\arguments{
\item{experimentName}{- an S4 object of class tssObject that contains information about the experiment}
}
\value{
\emph{importBam} fills the slot experimentName@bamData in the tssObject \emph{experimentName} with
        GAlignments objects from the \bold{GenomicAlignments} package, one for each parsed input
        BAM file.
}
\description{
\emph{importBam} processes BAM files as specified by \emph{initializeExp}. Files will be imported in the order that they appear on the @fileNames slot, which is in ascending alphanumeric order.
}

