% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeExp.R
\name{initializeExp}
\alias{initializeExp}
\title{Initializes the TSS profiling experiment}
\usage{
initializeExp(expTitle, objName, expDir, isPairedEnd)
}
\arguments{
\item{expTitle}{a title for the experiment (character)}

\item{objName}{a name the object to be created in your working environment (character)}

\item{expDir}{a path to the directory with the alignment files in bam format (character)}

\item{isPairedEnd}{a logical indicating whether the TSS profiling experiment was paired-end (logical)}
}
\value{
Creates a tssExp object in the user's current workspace.
}
\description{
Initializes the TSS profiling experiment
}

