% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeExp.R
\name{initializeExp}
\alias{initializeExp}
\title{\strong{initializeExp}}
\usage{
initializeExp(expTitle, experimentName, isPairedEnd)
}
\arguments{
\item{expTitle}{A descriptive title for the experiment (character)}

\item{experimentName}{The name for the \emph{tssObject} to be created
(character).
Your choice for \emph{experimentName} must lack spaces.}

\item{isPairedEnd}{specifies whether the TSS profiling experiment is
paired-end (if TRUE) or single-end (if FALSE) (logical)}
}
\value{
a new \emph{tssObject} with name \emph{experimentName}
}
\description{
\code{initializeExp} creates an initialized tssObject
in the user's workspace
}

