% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeExp.R
\name{initializeExp}
\alias{initializeExp}
\title{\strong{initializeExp}}
\usage{
initializeExp(expTitle, experimentName, expDir, isPairedEnd)
}
\arguments{
\item{expTitle}{A descriptive title for the experiment (character)}

\item{experimentName}{name for the \emph{tssObject} to be created in your working environment (character). Your choice for \emph{experimentName} must lack spaces.}

\item{expDir}{path to the directory with the alignment files in .bam format (character). Note that all the paths to all files in \emph{expDir} with the extension .bam in \emph{expDir} will be imported with this function.}

\item{isPairedEnd}{specifies whether the TSS profiling experiment is paired-end (if TRUE) or single-end (if FALSE) (logical)}
}
\value{
Creates a new \emph{tssObject} with the name \emph{experimentName} that is written to the user's working environment.
}
\description{
\code{initializeExp} initializes the TSS profiling experiment
}
\note{
The example provided can be found in \emph{Example 1} from the vignette (/inst/doc/TSRchitect.Rmd)

Please note that all .bam files found in \emph{expDir} will be retrieved and written in ascending alphanumeric order to the \emph{@fileNames} slot on the \emph{tssObject} that is created.
}
\examples{
 initializeExp(expTitle ="Human RAMPAGE", experimentName="Hs_RAMPAGE", expDir="hsRAMPAGEbam/", isPairedEnd=TRUE)
}

