% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeExpr.R
\name{mergeExpr}
\alias{mergeExpr}
\title{Combines samples from two different tss experiments into a single GRanges object}
\usage{
mergeExpr(expName)
}
\arguments{
\item{expName}{- an S4 object of class tssExp that contains information about the experiment.}
}
\value{
expData datasets will be merged (according to the sampleIDs) and assigned to your tssExp object
}
\description{
Combines samples from two different tss experiments into a single GRanges object
}

