% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeSampleData.R
\name{mergeSampleData}
\alias{mergeSampleData}
\title{Combines samples from two different tss experiments into a single GRanges object}
\usage{
mergeSampleData(experimentName)
}
\arguments{
\item{experimentName}{- an S4 object of class tssObject that contains information about the experiment.}
}
\value{
tssCountData datasets will be merged (according to the sampleIDs) and assigned to your tssObject object
}
\description{
Combines samples from two different tss experiments into a single GRanges object
}

