% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeTSS.R
\name{mergeTSS}
\alias{mergeTSS}
\title{Combines samples from different tss experiments into a single GRanges object}
\usage{
mergeTSS(experimentName)
}
\arguments{
\item{experimentName}{- an S4 object of class tssObject that contains information about the experiment.}
}
\value{
merged tss profiling experiments according to their assigned sampleIDs to the tssObject
}
\description{
Combines samples from different tss experiments into a single GRanges object
}

