% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTSS.R
\name{processTSS}
\alias{processTSS}
\title{processTSS
Finds TSRs from a given sequence}
\usage{
processTSS(experimentName, parallel, tssSet, writeTable = FALSE)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject containing information in slot tssTagData}

\item{parallel}{- whether to run in parallel or not (logical)}

\item{tssSet}{- default is "all"; for specific use, specify tssSet number (as character)}

\item{writeTable}{- specifies whether the output should be written to a table. (logical)}
}
\value{
creates a list of GenomicRanges containing TSR positions in slot 'tsrData' on your tssObject object
}
\description{
processTSS
Finds TSRs from a given sequence
}

