% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTSS.R
\name{processTSS}
\alias{processTSS}
\title{\strong{processTSS}}
\usage{
processTSS(experimentName, parallel, tssSet, writeTable = FALSE)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject} containing information in slot \emph{@tssTagData}}

\item{parallel}{whether to run in parallel or not (logical)}

\item{tssSet}{default is "all"; for specific use, specify \emph{tssSet} number (as character)}

\item{writeTable}{specifies whether the output should be written to a table. (logical)}
}
\value{
creates a list of \linkS4class{GenomicRanges} containing TSS positions in slot \emph{tssTagData} on the \emph{tssObject}
}
\description{
\code{processTSS} calulates the number of observed reads at a given TSS coordinate across an entire dataset.
}
\note{
The example provided can be found in \emph{Example 1} from the vignette (/inst/doc/TSRchitect.Rmd)

Note that the \emph{tssSet} parameter must be of class \emph{character}, even when selecting an individual dataset.
}
\examples{
processTSS(experimentName=Hs_RAMPAGE, parallel=TRUE, tssSet="all", writeTable=TRUE)
}

