% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTSS.R
\name{processTSS}
\alias{processTSS}
\title{processTSS
Creates an expression matrix for all TSSs within a given TSS experiment (in tssData)}
\usage{
processTSS(experimentName, tssSet, writeTable)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject containing information in slot tssData}

\item{tssSet}{- number of the dataset to be analyzed}

\item{writeTable}{if TRUE, writes a data frame containing the TSSs positions and their abundance to your workspace}
}
\value{
creates a data frame containing tss expression for each TSS position in slot 'expData' on your tssObject object
}
\description{
processTSS
Creates an expression matrix for all TSSs within a given TSS experiment (in tssData)
}

