% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processTSS.R
\name{processTSS}
\alias{processTSS}
\title{\strong{processTSS}}
\usage{
processTSS(experimentName, parallel, n.cores, tssSet, writeTable)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject} containing
information in slot \emph{@tssTagData}}

\item{parallel}{if TRUE, the job will be run in parallel (logical)}

\item{n.cores}{the number of cores to be used for this job.
Ignored if 'parallel=FALSE' (numeric)}

\item{tssSet}{default is "all"; to select a single \emph{tssSet},
specify it (as character)}

\item{writeTable}{specifies whether the output should be written
to a file. (logical)}
}
\value{
Creates a list of \linkS4class{GenomicRanges} containing TSS
positions in slot \emph{tssTagData} on the \emph{tssObject}.
}
\description{
\code{processTSS} calulates the number of observed reads
at a given TSS coordinate across an entire dataset.
}
\note{
Note that the \emph{tssSet} parameter must be of class
\emph{character}, even when selecting an individual dataset.

An example similar to the one provided can be found in
the vignette (/inst/doc/TSRchitect.Rmd).
}
\examples{
load(system.file("extdata", "tssObjectExample.RData",
package="TSRchitect"))
processTSS(experimentName=tssObjectExample, parallel=FALSE, tssSet="all",
writeTable=FALSE)

}

