% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr-metrics.R
\name{shapeIndex}
\alias{shapeIndex}
\title{shapeIndex}
\usage{
shapeIndex(x)
}
\value{
Calculates the shape index (SI) for a given TSR
}
\description{
An internal function that caculates the shape index (SI)
of a given TSR from the output of tsrCluster
}
\keyword{internal}

