% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tagCountTSS}
\alias{tagCountTSS}
\title{tagCountTSS}
\usage{
tagCountTSS(y, outfname = "TSS.txt", writeDF = FALSE)
}
\value{
a matrix [a, h] containing the number of unique TSSs (a) and their
abundances (h).
}
\description{
an internal function that returns a matrix [a, h] where a = the
number of unique TSSs and h = the # of tags observed at that position
}
\keyword{internal}
