% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tagCountTSS}
\alias{tagCountTSS}
\title{tagCountTSS}
\usage{
tagCountTSS(y, n.cores = 1, outfname = "TSS.txt", writeTable = FALSE)
}
\arguments{
\item{y}{a data frame containing the contents of a single slot of tssTagData.}

\item{n.cores}{the number of cores to be used for this job.}

\item{outfname}{the prefix of the file name of TSS information to be written.
(character)}

\item{writeTable}{if TRUE, the tag count information is written to a file in the
workding directory (logical)}
}
\value{
a matrix [a, h] containing the number of unique TSSs (a) and
their corresponding abundances (h) is returned.
}
\description{
an internal function that returns a matrix [a, h] where a = the
number of unique TSSs and h = the # of tags observed at that position.
tagCountTSS is invoked via prcTSS.
}
\keyword{internal}
