% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSRchitectUsersGuide.R
\name{TSRchitectUsersGuide}
\alias{TSRchitectUsersGuide}
\title{\strong{TSRchitectUsersGuide}}
\usage{
TSRchitectUsersGuide(view = TRUE)
}
\arguments{
\item{view}{(logical) if TRUE (default) the User's Guide
is opened in the local pdf viewer.
If FALSE then the full path to the User's Guide is
returned.}
}
\value{
if view=FALSE, then the full path to the User's Guide
is returned.
}
\description{
Opens the TSRchitect User's Guide in a pdf viewer
on the user's system.
}
\examples{
myPath <- TSRchitectUsersGuide(view=FALSE)

}

