% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tsrCluster}
\alias{tsrCluster}
\title{tsrCluster}
\usage{
tsrCluster(x, minNbrTSSs = 3, minDist = 20)
}
\value{
A data frame of TSRs with variables\cr
\enumerate{
        \item seq = sequence identifier (seq)
        \item start = start of TSR (num)
        \item end = end of TSR (num)
        \item strand = + or - (factor)
        \item nTSSs = count of TSSs (num)
        \item tsrWidth = width of TSR (num)
        \item shapeIndex = shape index value of TSR (num)
}
}
\description{
Partitions, then clusters TSS data by sequence to create
a data frame of identified TSR positions and other associated metrics.
}
\keyword{internal}
