% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tsrCluster}
\alias{tsrCluster}
\title{tsrCluster}
\usage{
tsrCluster(x, minNbrTSSs = 3, minDist = 20)
}
\value{
A list of TSRs from the data frame generated by
\code{\link{tagCountTSS}}
}
\description{
Partitions, then clusters tss data by sequence.
}
\keyword{internal}
