% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr-metrics.R
\name{tsrCounts}
\alias{tsrCounts}
\title{An internal function that caculates the total number of counts associated with a given TSR}
\usage{
tsrCounts(x)
}
\description{
An internal function that caculates the total number of counts associated with a given TSR
}

