% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr-metrics.R
\name{tsrCounts}
\alias{tsrCounts}
\title{tsrCounts}
\usage{
tsrCounts(x)
}
\value{
Returns the number of TSS tags associated with a TSR
}
\description{
An internal function that caculates the total number of
counts associated with a given TSR
}
\keyword{internal}
