% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrFind.R
\name{tsrFind}
\alias{tsrFind}
\title{tsrFind
Finds TSRs from a given chromosome}
\usage{
tsrFind(expName, tssNum, nTSSs, clustDist)
}
\arguments{
\item{expName}{an object of tssExp format containing information in slot tssData}

\item{tssNum}{the number of the dataset to be analyzed}
}
\value{
creates a list of GenomicRanges containing TSR positions in slot 'tsrData' on your tssExp object
}
\description{
tsrFind
Finds TSRs from a given chromosome
}

