% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrFind.R
\name{tsrFind}
\alias{tsrFind}
\title{tsrFind
Finds TSRs from a given sequence}
\usage{
tsrFind(experimentName, tssSet = 1, tagCountThreshold, clustDist,
  setToCluster, writeTable = FALSE)
}
\arguments{
\item{experimentName}{- a S4 object of class tssObject containing information in slot tssTagData}

\item{tssSet}{- number of the dataset to be analyzed}

\item{tagCountThreshold}{- number of TSSs required at a given position}

\item{clustDist}{- maximum distance of TSSs between two TSRs (in base pairs)}

\item{setToCluster}{- specifies the set to be clustered. Options are "replicates" or "merged".}

\item{writeTable}{- specifies whether the output should be written to a table. (logical)}
}
\value{
creates a list of GenomicRanges containing TSR positions in slot 'tsrData' on your tssObject object
}
\description{
tsrFind
Finds TSRs from a given sequence
}

