% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{TSRshapeIndex}
\alias{TSRshapeIndex}
\title{TSRshapeIndex}
\usage{
TSRshapeIndex(tssArray)
}
\arguments{
\item{tssArray}{an object containing TSS coordinates and their
abundances. (data.frame)}
}
\value{
Returns a  numeric vector of length 1 containing the Shape Index (SI)
value for the selected TSR.
}
\description{
An internal function that caculates the shape index (SI)
of a given TSR from the output of tsrCluster.
TSRshapeIndex is called by tssArrayProperties().
}
\keyword{internal}
