% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrToAnnotation.R
\name{tsrToAnnotation}
\alias{tsrToAnnotation}
\title{tsrToAnnotation}
\usage{
tsrToAnnotation(experimentName, upstreamDist)
}
\arguments{
\item{experimentName}{- an S4 object of class tssObject that contains information about the experiment.}

\item{upstreamDist}{- the maximum distance (in bp) upstream of the selected interval necessary to associate a TSR with a given annotation.}
}
\value{
tsrToAnnotation adds geneID information to the tsrData data frame and attaches it to the tssObject
experimentName
}
\description{
tsrToAnnotation associates an identified promoter with a given gene, if found upstream and on the same strand within a selected distance.
}

