% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsrToDF.R
\name{tsrToDF}
\alias{tsrToDF}
\title{\emph{tsrToDF()}}
\usage{
tsrToDF(x)
}
\arguments{
\item{x}{A list with TSR data as returned by tsrCluster()}
}
\value{
A data frame of TSRs with variables\cr
\enumerate{
        \item seq = sequence identifier (seq)
        \item start = start of TSR (num)
        \item end = end of TSR (num)
        \item strand = + or - (factor)
        \item nTSSs = count of TSSs (num)
        \item tsrWidth = width of TSR (num)
        \item shapeIndex = shape index value of TSR (num)
}
}
\description{
\emph{tsrToDF} is an internal utility function in
that converts a \emph{tsrCluster()}-generated list of TSR data to
a data frame.
}
\keyword{internal}
