% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsr-metrics.R
\name{tsrWidth}
\alias{tsrWidth}
\title{tsrWidth}
\usage{
tsrWidth(x)
}
\value{
Returns a width value (in bp) for a given TSR
}
\description{
An internal function that caculates the width
of a given TSR from the output of tsrCluster
}
\keyword{internal}
