% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tssArrayProperties}
\alias{tssArrayProperties}
\title{tssArrayProperties}
\usage{
tssArrayProperties(tssArray, seqName, strand)
}
\value{
A vector describing the TSR
}
\description{
An internal function that caculates various properties
for a TSR derived in tsrCluster()
}
\keyword{internal}
