% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tssChr}
\alias{tssChr}
\title{tssChr (internal function)
Retreives tss data from a given experiment by chromosome.}
\usage{
tssChr(expName, tssNum, chrName)
}
\arguments{
\item{expName}{an object of class tssExp containing the tss data}

\item{tssNum}{the slot number of the tss data to be clustered}

\item{chrName}{the name of the chromosome to select}
}
\value{
a list object containing TSS data for a single chromosome (plus and minus strands)
}
\description{
tssChr (internal function)
Retreives tss data from a given experiment by chromosome.
}

