% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tssChr}
\alias{tssChr}
\title{tssChr (internal function)
Retreives tss data from a given experiment by sequence.}
\usage{
tssChr(tssObj, seqName)
}
\arguments{
\item{tssObj}{an object of class GRanges containing data from a slot of tssTagData}

\item{seqName}{the name of the sequence to select}
}
\value{
a list object containing TSS data for a single sequence (plus and minus strands)
}
\description{
tssChr (internal function)
Retreives tss data from a given experiment by sequence.
}

