% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering-functions.R
\name{tssChr}
\alias{tssChr}
\title{tssChr (internal function)
Retreives tss data from a given experiment by chromosome.}
\usage{
tssChr(tssObj, chrName)
}
\arguments{
\item{tssObj}{an object of class GRanges containing data from a slot of tssData}

\item{chrName}{the name of the chromosome to select}
}
\value{
a list object containing TSS data for a single chromosome (plus and minus strands)
}
\description{
tssChr (internal function)
Retreives tss data from a given experiment by chromosome.
}

