% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tssToBed.R
\name{tssToBed}
\alias{tssToBed}
\title{Converts TSS data into bed format}
\usage{
tssToBed(expName, fileName = "myTSS")
}
\arguments{
\item{expName}{an object of class tssExp with tss data loaded}

\item{fileName}{the name of the BED file to be written}
}
\description{
Converts TSS data into bed format
}

