% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeTSS.R
\name{writeTSS}
\alias{writeTSS}
\alias{writeTSS,tssObject,character,numeric,character,logical,character-method}
\title{\strong{writeTSS}}
\usage{
writeTSS(
  experimentName,
  tssSetType,
  tssSet = 1,
  tssLabel = "TSS_",
  mixedorder = FALSE,
  fileType = "tab"
)

\S4method{writeTSS}{tssObject,character,numeric,character,logical,character}(
  experimentName,
  tssSetType,
  tssSet = 1,
  tssLabel = "TSS_",
  mixedorder = FALSE,
  fileType = "tab"
)
}
\arguments{
\item{experimentName}{an S4 object of class \emph{tssObject} containing
information in slot \emph{@tssCountData}}

\item{tssSetType}{specifies the set to be written to file.
Options are "replicates" or "merged". (character)}

\item{tssSet}{number of the dataset to be processed (numeric).}

\item{tssLabel}{specifies the label to be used in the name column of
BED format output}

\item{mixedorder}{a logical specifying whether the sequence names should
be ordered alphanumerically ("10" following "9" rather than "1"). (logical)}

\item{fileType}{the format of the file to be written.
Possible choices are "tab" for tab-delimited output, "bed" for BED format,
and "bedGraph" for BedGraph format (character). In case "bedGraph," three
output files are produced: *_PLUS.bedGraph and *_MINUS.bedGraph, which
record the number of TSS tags per position on plus and minus strand
separately; and a combined *.bedGraph file which lists plus strand tag
counts as positive numbers, minus strand tag counts as negative numbers,
and positions with both plus and minus strand tag counts display the
tag count corresponding to the highest absolute value.}
}
\value{
A table containing the specified TSS data set that
is to be written to your working directory.
}
\description{
\code{writeTSS} writes identified TSSs
from a specified data set to a file in either tab or BED formats
}
\note{
The .bed file written adheres to the standard six-column BED format,
while "tab" format is identical to that of the data.frames containing TSS
data.

For more information on the BED format, please visit
https://genome.ucsc.edu/FAQ/FAQformat#format1
}
\examples{
load(system.file("extdata", "tssObjectExample.RData", package="TSRchitect"))
writeTSS(experimentName=tssObjectExample, tssSetType="replicates",
         tssSet=1, tssLabel="TSSsample1_", mixedorder=FALSE, fileType="tab")

}
